Add files via upload
Added linked list structure, to avoid memory errors/

+++ RandomFunctions.h
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <math.h>
 #include <time.h>
+#include <stdint.h>
 
 #define IS_EQUAL(x, m) x >= 0 ? x : x + m
 
@@ -12,4 +13,4 @@ int64_t mod(int64_t a, int64_t b);
 
 int is_simple(unsigned long long x1);
 void print_byte(int32_t var, int size);
-int getAbsStartRand();
+int64_t getAbsStartRand();
